CRAN required changes.
From: Kurt Hornik <Kurt.Hornik@R-project.org>
Sent: Sunday, March 22, 2020 6:08 AM
To: Gallas, Brandon D. <Brandon.Gallas@fda.hhs.gov>
Cc: CRAN@R-project.org
Subject: CRAN package iMRMC

Dear maintainer,

Please see the problems shown on
<https://cran.r-project.org/web/checks/check_results_iMRMC.html>.

Specifically, see the problems shown for the r-devel checks.

These can be reproduced by checking with --as-cran using a very current r-devel (r77865 or later), which makes data.frame() and read.table() use a stringsAsFactors = FALSE default, which is planned to become the new default for the upcoming R 4.0.0.

Please see
<https://developer.r-project.org/Blog/public/2020/02/16/stringsasfactors/index.html>
for more information about this change.

Can you please fix your package to work with both the old and new default?  In principle, this can easily be achieved by adding stringsAsFactors = TRUE to the relevant calls to data.frame() or
read.table() [or other read.* function calling read.table()], but please only do this if the sort order used in the string to factor conversion really does not matter (see the blog post about the locale dependence of the conversion).  Otherwise, please change to create the factors with explicitly given levels.

The new problems may be from code in a package you depend on: in this case, please let me know, and get in touch with the maintainer of that package.

Please correct before 2020-04-05 to safely retain your package on CRAN.

Note that this will be the *final* reminder.